+++ src/Ferrite.jl
@@ -63,6 +63,7 @@ include(joinpath("Adaptivity", "AdaptiveCells.jl"))
 include("Dofs/DofHandler.jl")
 include("Dofs/MixedDofHandler.jl")
 include("Dofs/ConstraintHandler.jl")
+include("Dofs/apply_analytical.jl")
 include("Dofs/DofRenumbering.jl")
 
 include("iterators.jl")