+++ src/arrayutils.jl
@@ -9,6 +9,7 @@ end
 
 """
     addindex!(A::AbstractMatrix{T}, v::T, i::Int, j::Int)
+    addindex!(b::AbstractVector{T}, v::T, i::Int)
 
 Equivalent to `A[i, j] += v` but more efficient.
 
@@ -18,12 +19,46 @@ This method avoids the double lookup.
 
 Zeros are ignored (i.e. if `iszero(v)`) by returning early. If the index `(i, j)` is not
 existing in the sparsity pattern of `A` this method throws a `SparsityError`.
-"""
-addindex!(A, v, i, j)
 
-function addindex!(A::SparseMatrixCSC{Tv}, v, i::Integer, j::Integer) where Tv
-    return addindex!(A, Tv(v), Int(i), Int(j))
+Fallback: `A[i, j] += v`.
+"""
+addindex!
+
+function addindex!(A::AbstractMatrix{T}, v, i::Integer, j::Integer) where T
+    return addindex!(A, T(v), Int(i), Int(j))
 end
+function addindex!(A::AbstractMatrix{T}, v::T, i::Int, j::Int) where T
+    iszero(v) && return A
+    A[i, j] += v
+    return A
+end
+
+function addindex!(b::AbstractVector{T}, v, i::Integer) where T
+    return addindex!(b, T(v), Int(i))
+end
+function addindex!(b::AbstractVector{T}, v::T, i::Int) where T
+    iszero(v) && return b
+    b[i] += v
+    return b
+end
+
+"""
+    fillzero!(A::AbstractVecOrMat{T})
+
+Fill the (stored) entries of the vector or matrix `A` with zeros.
+
+Fallback: `fill!(A, zero(T))`.
+"""
+fillzero!(A)
+
+function fillzero!(A::AbstractVecOrMat{T}) where T
+    return fill!(A, zero(T))
+end
+
+##################################
+## SparseArrays.SparseMatrixCSC ##
+##################################
+
 function addindex!(A::SparseMatrixCSC{Tv}, v::Tv, i::Int, j::Int) where Tv
     @boundscheck checkbounds(A, i, j)
     # Return early if v is 0
@@ -40,4 +75,27 @@ function addindex!(A::SparseMatrixCSC{Tv}, v::Tv, i::Int, j::Int) where Tv
         # (i, j) not stored. Throw.
         throw(SparsityError())
     end
+end
+
+function fillzero!(A::SparseMatrixCSC{T}) where T
+    fill!(nonzeros(A), zero(T))
+    return A
+end
+function fillzero!(A::Symmetric{T,<:SparseMatrixCSC}) where T
+    fillzero!(A.data)
+    return A
+end
+
+# Compat and convenience layer around SparseArrays.spzeros! (SparseArrays.jl#284, SparseArrays.jl#315)
+function spzeros!!(::Type{Tv}, I::Vector{Ti}, J::Vector{Ti}, m::Int, n::Int) where {Tv, Ti <: Integer}
+    @assert length(I) == length(J)
+    @static if isdefined(SparseArrays, :spzeros!)
+        klasttouch = Vector{Ti}(undef, n)
+        csrrowptr  = Vector{Ti}(undef, m + 1)
+        csrcolval  = Vector{Ti}(undef, length(I))
+        S = SparseArrays.spzeros!(Tv, I, J, m, n, klasttouch, csrrowptr, csrcolval, I, J)
+    else
+        S = sparse(I, J, zeros(Tv, length(I)), m, n)
+    end
+    return S
 end